# Improve performance of PDF::Reader::Filter::Depredict#png_depredict