Calling Array#slice! for each row could impact performance. When pdf files have cross reference stream with predictor functions, it takes really long time to open them.