@yob 2018-07-22
Thanks for this patch, and apologies for the slow response. A new job and a new baby have kept me distracted.

Do you have any sample code that demonstrates the performance issue?

I tried to re-create it by extracting a stream that enters the `png_deprdict` method (from the `spec/data/type3_font.pdf` file), but the peformance with and without your patch was roughly comparable. I suspect my sample data is too small to trigger the issue.

```ruby 
require 'pdf/reader'
require 'allocation_tracer'
require 'pp'
require 'benchmark/ips'

ObjectSpace::AllocationTracer.setup(%i{type})

data = "\x02\x00\x00\x00\xFF\xFF\x02\x02\x00\f\x01\a\x02\x00\x00\x00\x00\xFB\x02\xFF\x00\x9D\x00\xFF\x02\x00\x00f\x00\x00\x02\x00\x01I\x00\x00\x02\x00\xFFc\x00\x00\x02\x00\x02I\x00\x00\x02\x00\xFFf\x00\x00\x02\x00\x01\a\x00\x00\x02\x00\x00\xA5\x00\x00\x02\x01\xFE\xF6\x00\x00\x02\xFF\x05Y\x00\x00\x02\x01\xFB\xA7\x00\x05\x02\xFF\x03\x06\x00\xFB\x02\x00\xFFq\x00\x00\x02\x01\xFE\x89\x00\x03\x02\x00\x00\x00\x00\xFF\x02\xFF\x03\x19\x00\xFE\x02\x01\xFD\xE7\x00\x04\x02\xFF\x03\xE8\x00\xFC\x02\x00\x02/\x00\x00\x02\x00\x00\x14\x00\x00\x02\x00\x02\xDE\x00\x00\x02\x00\x00\x14\x00\x00"

res = ObjectSpace::AllocationTracer.trace do
  filter = PDF::Reader::Filter::Depredict.new(Predictor: 12, Columns: 5)
  filter.filter(data)
end

pp res

filter = PDF::Reader::Filter::Depredict.new(Predictor: 12, Columns: 5)
Benchmark.ips do |x|
  x.time = 5
  x.warmup = 2
  x.report("depredict") { 
    filter.filter(data)
  }
end
```  

## master

```bash
$ ruby -Ilib bench-depredict.rb 
{[:T_HASH]=>[1, 0, 0, 0, 0, 0],
 [:T_OBJECT]=>[1, 0, 0, 0, 0, 0],
 [:T_ARRAY]=>[254, 0, 0, 0, 0, 0],
 [:T_IMEMO]=>[102, 0, 0, 0, 0, 0],
 [:T_STRING]=>[27, 0, 0, 0, 0, 0],
 [:T_DATA]=>[1, 0, 0, 0, 0, 0]}


Warming up --------------------------------------
           depredict     1.041k i/100ms
Calculating -------------------------------------
           depredict     10.037k (±10.0%) i/s -     49.968k in   5.052510s
```

## patched

```bash
$ ruby -Ilib bench-depredict.rb 
{[:T_HASH]=>[1, 0, 0, 0, 0, 0],
 [:T_OBJECT]=>[1, 0, 0, 0, 0, 0],
 [:T_ARRAY]=>[255, 0, 0, 0, 0, 0],
 [:T_IMEMO]=>[104, 0, 0, 0, 0, 0],
 [:T_STRING]=>[27, 0, 0, 0, 0, 0],
 [:T_DATA]=>[1, 0, 0, 0, 0, 0]}
Warming up --------------------------------------
           depredict     1.013k i/100ms
Calculating -------------------------------------
           depredict     10.141k (± 4.2%) i/s -     50.650k in   5.006079s
```